+++ clients/validator/src/stargate-helper.ts
@@ -7,13 +7,12 @@ const mainnetDenom = 'nym';
 export function nymGasPrice(prefix: string): GasPrice {
   if (typeof prefix === 'string') {
     if (prefix === mainnetPrefix) {
-      prefix = mainnetDenom;
+      return GasPrice.fromString(`0.025u${mainnetDenom}`);
     }
     return GasPrice.fromString(`0.025u${prefix}`); // TODO: ideally this ugly conversion shouldn't be hardcoded here.
   }
-  else {
-    throw new Error(`${prefix} is not of type string`);
-  }
+
+  throw new Error(`${prefix} is not of type string`);
 }
 
 export const downloadWasm = async (url: string): Promise<Uint8Array> => {